expoprts functions necessary for OSD interpolation from the window_function module
includes kepid as a part of the star object

includes kepid as part of a target object and determines how a kepid will be assigned

adjusts the calc_snr_if_transit function to work for osd instead of cdpp. Kept old version for testing purposes.

use newPDST.jl to determine if duration and period are compatible
calculate osd and use it to calculate snr
There is also functionality to save the ratio of new and old snr calculations in a global array, then plot a histogram of those ratios